requirements.txt
ctransformers==0.2.5
sentence-transformers==3.3.1
pinecone-client==5.0.1
langchain==0.3.14
langchain-huggingface==0.1.2
langchain-community
flask
pypdf
python-dotenv
-e .